# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Obss.Host | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5020/"
    }
  ],
  "paths": {
...
    "/api/v{version}/audit/entries": {
      "post": {
        "tags": [
          "Audit"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuditEntryCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "EntityType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "EntityId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "PerformedById",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "FromDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "ToDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v{version}/audit/entries/{id}": {
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v{version}/audit/entities/{entityType}/{entityId}": {
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "entityType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v{version}/audit/policies": {
      "post": {
        "tags": [
          "Audit"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuditPolicyCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v{version}/audit/sensitive-operations": {
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "From",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "To",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v{version}/audit/alerts": {
      "get": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "Severity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "AlertType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IsAcknowledged",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "FromDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "ToDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v{version}/audit/alerts/{id}/acknowledge": {
      "post": {
        "tags": [
          "Audit"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v{version}/audit/alert-rules": {
      "post": {
        "tags": [
          "Audit"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertRuleCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Audit"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v{version}/gateway/routes": {
      "post": {
        "tags": [
          "ApiGateway"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterApiRouteCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "ApiGateway"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v{version}/gateway/api-keys": {
      "post": {
        "tags": [
          "ApiGateway"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "ApiGateway"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v{version}/gateway/api-keys/{id}/revoke": {
      "post": {
        "tags": [
          "ApiGateway"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v{version}/gateway/partners": {
      "post": {
        "tags": [
          "ApiGateway"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPartnerCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "ApiGateway"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateAlertRuleCommand": {
        "required": [
          "name",
          "description",
          "alertType",
          "severity",
          "threshold",
          "windowMinutes"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "alertType": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "threshold": {
            "type": "integer",
            "format": "int32"
          },
          "windowMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "CreateApiKeyCommand": {
        "required": [
          "name",
          "permissions",
          "allowedIPs",
          "rateLimitPerMinute",
          "partnerId",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "allowedIPs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "rateLimitPerMinute": {
            "type": "integer",
            "format": "int32"
          },
          "partnerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "CreateAuditEntryCommand": {
        "required": [
          "entityType",
          "entityId",
          "action",
          "changes",
          "performedById",
          "performedByName",
          "ipAddress",
          "userAgent",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "changes": {
            "type": "string",
            "nullable": true
          },
          "performedById": {
            "type": "string",
            "nullable": true
          },
          "performedByName": {
            "type": "string",
            "nullable": true
          },
          "ipAddress": {
            "type": "string",
            "nullable": true
          },
          "userAgent": {
            "type": "string",
            "nullable": true
          },
          "correlationId": {
            "type": "string",
            "nullable": true
          },
          "isSensitive": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "CreateAuditPolicyCommand": {
        "required": [
          "entityType",
          "retentionDays"
        ],
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string"
          },
          "retentionDays": {
            "type": "integer",
            "format": "int32"
          },
          "alertOnFailure": {
            "type": "boolean",
            "default": false
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "RegisterApiRouteCommand": {
        "required": [
          "path",
          "method",
          "targetModule",
          "targetPath",
          "requireAuthentication",
          "requiredPermissions",
          "rateLimitPerMinute"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "targetModule": {
            "type": "string"
          },
          "targetPath": {
            "type": "string"
          },
          "requireAuthentication": {
            "type": "boolean"
          },
          "requiredPermissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "rateLimitPerMinute": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "RegisterPartnerCommand": {
        "required": [
          "name",
          "contactName",
          "contactEmail",
          "allowedIPs",
          "slaLevel",
          "maxRequestsPerDay"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "contactName": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "allowedIPs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "slaLevel": {
            "$ref": "#/components/schemas/SlaLevel"
          },
          "maxRequestsPerDay": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "SlaLevel": {
        "enum": [
          "Basic",
          "Standard",
          "Premium"
        ]
      },
...
    }
  },
...
}